refactor: extract shared RelatedCoursesList component
